+++ app/__init__.py
@@ -39,7 +39,7 @@ def register_blueprint():
     from app.authentication.rest import auth_blueprint
     from app.events.rest import events_blueprint
     from app.fees.rest import fees_blueprint, fee_blueprint
-    from app.event_types.rest import event_types_blueprint, event_type_blueprint
+    from app.routes.event_types.rest import event_types_blueprint, event_type_blueprint
     from app.routes.speakers.rest import speakers_blueprint, speaker_blueprint
     application.register_blueprint(base_blueprint)
     application.register_blueprint(auth_blueprint)
